requirements.txt
setuptools~=57.4.0
requests~=2.26.0
enlighten~=1.10.1
Pillow~=8.4.0